+++ tests/test-sitemap-cpt.php
@@ -5,7 +5,7 @@
  * @package Metro_Sitemap/unit_tests
  */
 
-require_once( 'msm-sitemap-test.php' );
+require_once('TestCase.php');
 
 /**
  * Unit Tests to confirm Sitemaps are properly handled for Custom Post Types
