+++ tests/test-sitemap-index.php
@@ -5,7 +5,7 @@
  * @package Metro_Sitemap/unit_tests
  */
 
-require_once( 'msm-sitemap-test.php' );
+require_once('TestCase.php');
 
 /**
  * Unit Tests to validate indexes are properly generated by Metro_Sitemap
