+++ tests/test-sitemap-stats.php
@@ -5,7 +5,7 @@
  * @package Metro_Sitemap/unit_tests
  */
 
-require_once( 'msm-sitemap-test.php' );
+require_once('TestCase.php');
 
 /**
  * Unit Tests to validate stats are propelry generated by Metro_Sitemap
